Net 8 windows

+++ CameraControl.Devices/Nikon/NikonD4.cs
@@ -28,7 +28,6 @@
 
 #region
 
-using System;
 using CameraControl.Devices.Classes;
 
 #endregion
@@ -55,9 +54,9 @@ namespace CameraControl.Devices.Nikon
         {
             DeviceReady();
             StillImageDevice.ExecuteWithNoData(CONST_CMD_ChangeCameraMode, 1);
-            SetProperty(CONST_CMD_SetDevicePropValue, BitConverter.GetBytes((UInt16) 0x0001),
+            SetProperty(CONST_CMD_SetDevicePropValue, BitConverter.GetBytes((UInt16)0x0001),
                         CONST_PROP_ExposureProgramMode);
-            SetProperty(CONST_CMD_SetDevicePropValue, BitConverter.GetBytes((UInt32) 0xFFFFFFFF),
+            SetProperty(CONST_CMD_SetDevicePropValue, BitConverter.GetBytes(0xFFFFFFFF),
                         CONST_PROP_ExposureTime);
 
             ErrorCodes.GetException(CaptureInSdRam
@@ -93,27 +92,27 @@ namespace CameraControl.Devices.Nikon
         protected override PropertyValue<long> InitExposureDelay()
         {
             PropertyValue<long> res = new PropertyValue<long>()
-                                          {Name = "Exposure delay mode", IsEnabled = true, Code = 0xD06A};
+            { Name = "Exposure delay mode", IsEnabled = true, Code = 0xD06A };
             res.AddValues("3 sec", 0);
             res.AddValues("2 sec", 1);
             res.AddValues("One sec", 1);
             res.AddValues("OFF", 1);
             res.ReloadValues();
             res.ValueChanged +=
-                (sender, key, val) => SetProperty(CONST_CMD_SetDevicePropValue, new[] {(byte) val}, res.Code);
+                (sender, key, val) => SetProperty(CONST_CMD_SetDevicePropValue, new[] { (byte)val }, res.Code);
             return res;
         }
 
         public override void StartRecordMovie()
         {
-            SetProperty(CONST_CMD_SetDevicePropValue, new[] {(byte) 1}, CONST_PROP_ApplicationMode);
+            SetProperty(CONST_CMD_SetDevicePropValue, new[] { (byte)1 }, CONST_PROP_ApplicationMode);
             base.StartRecordMovie();
         }
 
         public override void StopRecordMovie()
         {
             base.StopRecordMovie();
-            SetProperty(CONST_CMD_SetDevicePropValue, new[] {(byte) 0}, CONST_PROP_ApplicationMode);
+            SetProperty(CONST_CMD_SetDevicePropValue, new[] { (byte)0 }, CONST_PROP_ApplicationMode);
         }
     }
 }
